chore: add dockerfile for example apps (#530)
Instead of running the examples apps directly on the server i use a dockerimage which is automatically created, pushed to dockerhub and deployed to the server. So the server will always have the newest version.
Additionally i will check the livedemo twice a day and create a issue if its down.